add refactor to extract post comment to its own component
